+++ src/IntermixBase.hpp
@@ -9,6 +9,7 @@ struct IntermixBase {
     typedef float (*IntermixMatrix)[PORTS];
 	virtual IntermixMatrix expGetCurrentMatrix() { return NULL; }
 	virtual int expGetChannelCount() { return 0; }
+	virtual void expSetFade(int i, float* fadeIn, float* fadeOut) { }
 };
 
 } // namespace Intermix