+++ src/plugin.hpp
@@ -35,6 +35,7 @@ extern Model* modelHive;
 extern Model* modelIntermix;
 extern Model* modelIntermixGate;
 extern Model* modelIntermixEnv;
+extern Model* modelIntermixFade;
 extern Model* modelSail;
 extern Model* modelPile;
 extern Model* modelPilePoly;